Refactor views into base views
